clips-executive: keep sync map entries on deletes
Do not delete the sync map entry when deleting a fact (either by a
trigger or by a local fact removal). If we delete the sycnc map entry,
we no longer no when the deletion happened. If we then receive a fact
insertion, we don't know whether that insertion was before or after the
deletion. Thus, keep the sync map entry, but set the fact idx to 0 and
update the timestamp.