clips-executive: add sync map entry on delete if necessary
If we have never seen a fact id but receive a delete trigger for that
id, assert a sync map entry so we can keep track of the deletion
timestamp. This is necessary because if another entity adds a fact and
then deletes the fact again, we cannot guarantee that we receive the
triggers for those events in the right order. If the order is reversed
and we have never seen the id before, we need to keep track when the
deletion happened, so we can skip the subsequent insertion trigger with
the older timestamp.

Co-Authored-By: Tarik Viehmann <tarik.viehmann@rwth-aachen.de>
Co-Authored-By: Daniel Habering <daniel.habering@rwth-aachen.de>